Change image-missing-symbolic to better alternatives
This should hopefully make image loading less distractive, as it's just an image appearing compared to an image switching out. This follows the pattern most browsers handle it. It should also not cause any layout shifts.

I also changed out the symbolic icon of the update and install pages to be a fully colored alternative, making it a better fit with other entries on that page that do have colored icons.